requirements.txt
pandas
numpy
sklearn
plotly.express
datetime
matplotlib
seaborn